Version Aug 14th
The outside nodes are represented as sub trees of inner side nodes.
When outside node is clicked, the window will shift to position that the
clicked node is inner side node.

To change files, you can modify the parameter of d3.json function.
There are two complex examples and six simple examples.